- move initializer to own class to make lint happy
