+++ community-material-typeface-library/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
             android:exported="false"
             tools:node="merge">
             <meta-data
-                android:name="com.mikepenz.iconics.typeface.library.community.material.CommunityMaterial$Initializer"
+                android:name="com.mikepenz.iconics.typeface.library.community.material.Initializer"
                 android:value="androidx.startup" />
         </provider>
     </application>
