+++ community-material-typeface-library/src/main/java/com/mikepenz/iconics/typeface/library/community/material/CommunityMaterial.kt
@@ -15,11 +15,8 @@
  */
 package com.mikepenz.iconics.typeface.library.community.material
 
-import android.content.Context
 import com.mikepenz.iconics.typeface.IIcon
 import com.mikepenz.iconics.typeface.ITypeface
-import com.mikepenz.iconics.typeface.IconicsHolder
-import com.mikepenz.iconics.typeface.IconicsInitializer
 import com.mikepenz.iconics.typeface.library.community.R
 import java.util.LinkedList
 
@@ -81,17 +78,6 @@ object CommunityMaterial : ITypeface {
             // ignore error, if not in 1st set, it has to be in the second
         }
         return Icon3.valueOf(key)
-    }
-
-    class Initializer : androidx.startup.Initializer<ITypeface> {
-        override fun create(context: Context): ITypeface {
-            IconicsHolder.registerFont(CommunityMaterial)
-            return CommunityMaterial
-        }
-
-        override fun dependencies(): List<Class<out androidx.startup.Initializer<*>>> {
-            return listOf(IconicsInitializer::class.java)
-        }
     }
 
     enum class Icon constructor(override val character: Char) : IIcon {
