+++ devicon-typeface-library/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
             android:exported="false"
             tools:node="merge">
             <meta-data
-                android:name="com.mikepenz.iconics.typeface.library.devicon.DevIcon$Initializer"
+                android:name="com.mikepenz.iconics.typeface.library.devicon.Initializer"
                 android:value="androidx.startup" />
         </provider>
     </application>
