+++ entypo-typeface-library/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
             android:exported="false"
             tools:node="merge">
             <meta-data
-                android:name="com.mikepenz.iconics.typeface.library.entypo.Entypo$Initializer"
+                android:name="com.mikepenz.iconics.typeface.library.entypo.Initializer"
                 android:value="androidx.startup" />
         </provider>
     </application>
