+++ material-design-dx-typeface-library/src/main/java/com/mikepenz/iconics/typeface/library/materialdesigndx/MaterialDesignDx.kt
@@ -15,11 +15,8 @@
  */
 package com.mikepenz.iconics.typeface.library.materialdesigndx
 
-import android.content.Context
 import com.mikepenz.iconics.typeface.IIcon
 import com.mikepenz.iconics.typeface.ITypeface
-import com.mikepenz.iconics.typeface.IconicsHolder
-import com.mikepenz.iconics.typeface.IconicsInitializer
 import java.util.LinkedList
 
 @Suppress("EnumEntryName")
@@ -63,17 +60,6 @@ object MaterialDesignDx : ITypeface {
         get() = "https://www.apache.org/licenses/LICENSE-2.0"
 
     override fun getIcon(key: String): IIcon = Icon.valueOf(key)
-
-    class Initializer : androidx.startup.Initializer<ITypeface> {
-        override fun create(context: Context): ITypeface {
-            IconicsHolder.registerFont(MaterialDesignDx)
-            return MaterialDesignDx
-        }
-
-        override fun dependencies(): List<Class<out androidx.startup.Initializer<*>>> {
-            return listOf(IconicsInitializer::class.java)
-        }
-    }
 
     enum class Icon constructor(override val character: Char) : IIcon {
         gmf_10k('\ue951'),
