+++ pixeden-7-stroke-typeface-library/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
             android:exported="false"
             tools:node="merge">
             <meta-data
-                android:name="com.mikepenz.iconics.typeface.library.pixeden7stroke.Pixeden7Stroke$Initializer"
+                android:name="com.mikepenz.iconics.typeface.library.pixeden7stroke.Initializer"
                 android:value="androidx.startup" />
         </provider>
     </application>
