+++ pixeden-7-stroke-typeface-library/src/main/java/com/mikepenz/iconics/typeface/library/pixeden7stroke/Pixeden7Stroke.kt
@@ -15,11 +15,8 @@
  */
 package com.mikepenz.iconics.typeface.library.pixeden7stroke
 
-import android.content.Context
 import com.mikepenz.iconics.typeface.IIcon
 import com.mikepenz.iconics.typeface.ITypeface
-import com.mikepenz.iconics.typeface.IconicsHolder
-import com.mikepenz.iconics.typeface.IconicsInitializer
 import java.util.LinkedList
 
 @Suppress("EnumEntryName")
@@ -63,17 +60,6 @@ object Pixeden7Stroke : ITypeface {
         get() = "http://themes-pixeden.com/font-demos/7-stroke/"
 
     override fun getIcon(key: String): IIcon = Icon.valueOf(key)
-
-    class Initializer : androidx.startup.Initializer<ITypeface> {
-        override fun create(context: Context): ITypeface {
-            IconicsHolder.registerFont(Pixeden7Stroke)
-            return Pixeden7Stroke
-        }
-
-        override fun dependencies(): List<Class<out androidx.startup.Initializer<*>>> {
-            return listOf(IconicsInitializer::class.java)
-        }
-    }
 
     enum class Icon constructor(override val character: Char) : IIcon {
         pe7_7s_album('\ue6aa'),
