+++ typeicons-typeface-library/src/main/java/com/mikepenz/iconics/typeface/library/typeicons/Initializer.kt
@@ -1,0 +1,17 @@
+package com.mikepenz.iconics.typeface.library.typeicons
+
+import android.content.Context
+import com.mikepenz.iconics.typeface.ITypeface
+import com.mikepenz.iconics.typeface.IconicsHolder
+import com.mikepenz.iconics.typeface.IconicsInitializer
+
+class Initializer : androidx.startup.Initializer<ITypeface> {
+    override fun create(context: Context): ITypeface {
+        IconicsHolder.registerFont(Typeicons)
+        return Typeicons
+    }
+
+    override fun dependencies(): List<Class<out androidx.startup.Initializer<*>>> {
+        return listOf(IconicsInitializer::class.java)
+    }
+}
