+++ typeicons-typeface-library/src/main/java/com/mikepenz/iconics/typeface/library/typeicons/Typeicons.kt
@@ -15,11 +15,8 @@
  */
 package com.mikepenz.iconics.typeface.library.typeicons
 
-import android.content.Context
 import com.mikepenz.iconics.typeface.IIcon
 import com.mikepenz.iconics.typeface.ITypeface
-import com.mikepenz.iconics.typeface.IconicsHolder
-import com.mikepenz.iconics.typeface.IconicsInitializer
 import java.util.LinkedList
 
 @Suppress("EnumEntryName")
@@ -63,17 +60,6 @@ object Typeicons : ITypeface {
         get() = "http://scripts.sil.org/cms/scripts/page.php?item_id=OFL_web"
 
     override fun getIcon(key: String): IIcon = Icon.valueOf(key)
-
-    class Initializer : androidx.startup.Initializer<ITypeface> {
-        override fun create(context: Context): ITypeface {
-            IconicsHolder.registerFont(Typeicons)
-            return Typeicons
-        }
-
-        override fun dependencies(): List<Class<out androidx.startup.Initializer<*>>> {
-            return listOf(IconicsInitializer::class.java)
-        }
-    }
 
     enum class Icon constructor(override val character: Char) : IIcon {
         typ_adjust_brightness('\ue000'),
